requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
black==23.3.0
boto==2.49.0
boto3==1.28.63
botocore==1.31.63
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==3.4.8
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
exceptiongroup==1.1.2
fastapi==0.75.0
fastapi-jwt-auth==0.5.0
h11==0.13.0
httpcore==0.18.0
httptools==0.2.0
httpx==0.25.0
idna==3.3
iniconfig==2.0.0
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==1.0.1
MarkupSafe==2.1.1
motor==2.5.1
mypy-extensions==1.0.0
orjson==3.6.7
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.2.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==1.7.1
pymongo==3.12.3
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.5
PyYAML
requests==2.27.1
rsa==4.8
s3transfer==0.7.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
tornado==6.1
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.9
uvicorn==0.17.5
uvloop==0.17.0
watchgod==0.8.1
websockets==10.2
aioredis
